Little refactoring of the same parts;
