+++ src/main/kotlin/org/jetbrains/research/libsl/visitors/ActionVisitor.kt
@@ -6,7 +6,6 @@ import org.jetbrains.research.libsl.context.LslGlobalContext
 import org.jetbrains.research.libsl.nodes.ActionArgumentDescriptor
 import org.jetbrains.research.libsl.nodes.ActionDecl
 import org.jetbrains.research.libsl.type.GenericType
-import org.jetbrains.research.libsl.type.GenericTypeBound
 import org.jetbrains.research.libsl.utils.PositionGetter
 
 class ActionVisitor(
@@ -57,53 +56,5 @@ class ActionVisitor(
 
 
     private val LibSLParser.ActionDeclContext.actionGenerics: MutableList<GenericType>
-        get() = getGenerics(this.generic(), this.whereConstraints())
-
-    // TODO: union with getGenerics in FunctionVisitor.kt class
-    private fun getGenerics(
-        genericContext: LibSLParser.GenericContext,
-        whereContext: LibSLParser.WhereConstraintsContext
-    ): MutableList<GenericType> {
-
-        val actionGenerics: MutableList<GenericType> = mutableListOf()
-        // GenericType? - this is bad; Only temporary
-        val actionGenericsOrdered: LinkedHashMap<String, GenericType?> = linkedMapOf()
-
-        genericContext.typeArgument().forEach {
-            if (it.typeIdentifier() != null)
-                actionGenericsOrdered[it.typeIdentifier().name.text] = null
-            else
-                actionGenericsOrdered[it.typeIdentifierBounded().typeIdentifier().name.text] = null
-        }
-
-        for (typeConstraint in whereContext.typeConstraint()) {
-            val paramName = typeConstraint.paramName.text
-            val bound =
-                if (typeConstraint.paramConstraint.typeIdentifierBounded() == null) GenericTypeBound.EMPTY else GenericTypeBound.fromString(
-                    typeConstraint.paramConstraint.typeIdentifierBounded().genericBound().text
-                )
-            val constraints = mutableListOf(
-                if (typeConstraint.paramConstraint.typeIdentifier() != null)
-                    processTypeIdentifier(typeConstraint.paramConstraint.typeIdentifier())
-                else
-                    processTypeIdentifier(typeConstraint.paramConstraint.typeIdentifierBounded().typeIdentifier())
-            )
-
-            if (
-                actionGenericsOrdered[paramName] == null
-            ) {
-                actionGenericsOrdered[paramName] = GenericType(
-                    paramName,
-                    typeBound = bound,
-                    constraints = constraints,
-                    context = actionContext
-                )
-            } else {
-                actionGenericsOrdered[paramName]?.constraints?.addAll(constraints)
-            }
-        }
-
-        actionGenericsOrdered.forEach { it.value?.let { it1 -> actionGenerics.add(it1) } }
-        return actionGenerics
-    }
+        get() = getGenericTypes(this.generic(), this.whereConstraints(), context)
 }
