+++ src/main/kotlin/org/jetbrains/research/libsl/visitors/LibSLParserVisitor.kt
@@ -10,10 +10,7 @@ import org.jetbrains.research.libsl.nodes.references.TypeReference
 import org.jetbrains.research.libsl.nodes.references.builders.AnnotationReferenceBuilder
 import org.jetbrains.research.libsl.nodes.references.builders.TypeReferenceBuilder
 import org.jetbrains.research.libsl.nodes.references.builders.TypeReferenceBuilder.getReference
-import org.jetbrains.research.libsl.type.ArrayType
-import org.jetbrains.research.libsl.type.GenericTypeBound
-import org.jetbrains.research.libsl.type.RealType
-import org.jetbrains.research.libsl.type.Type
+import org.jetbrains.research.libsl.type.*
 import org.jetbrains.research.libsl.utils.PositionGetter
 
 abstract class LibSLParserVisitor<T>(open val context: LslContextBase) : LibSLParserBaseVisitor<T>() {
@@ -145,5 +142,52 @@ abstract class LibSLParserVisitor<T>(open val context: LslContextBase) : LibSLPa
         }
 
         return namedArgs
+    }
+
+    fun getGenericTypes(
+        genericContext: LibSLParser.GenericContext,
+        whereContext: LibSLParser.WhereConstraintsContext,
+        context: LslContextBase
+    ): MutableList<GenericType> {
+
+        val funGenerics: MutableList<GenericType> = mutableListOf()
+        // GenericType? - this is bad; Only temporary
+        val functionGenericsOrdered: LinkedHashMap<String, GenericType?> = linkedMapOf()
+
+        genericContext.typeArgument().forEach {
+            if (it.typeIdentifier() != null)
+                functionGenericsOrdered[it.typeIdentifier().name.text] = null
+            else
+                functionGenericsOrdered[it.typeIdentifierBounded().typeIdentifier().name.text] = null
+        }
+
+        for (typeConstraint in whereContext.typeConstraint()) {
+            val paramName = typeConstraint.paramName.text
+            val bound =
+                if (typeConstraint.paramConstraint.typeIdentifierBounded() == null) GenericTypeBound.EMPTY else GenericTypeBound.fromString(
+                    typeConstraint.paramConstraint.typeIdentifierBounded().genericBound().text
+                )
+            val constraints: MutableList<TypeReference> = mutableListOf(
+                if (typeConstraint.paramConstraint.typeIdentifier() != null)
+                    processTypeIdentifier(typeConstraint.paramConstraint.typeIdentifier())
+                else
+                    processTypeIdentifier(typeConstraint.paramConstraint.typeIdentifierBounded().typeIdentifier())
+            )
+            if (
+                functionGenericsOrdered[paramName] == null
+            ) {
+                functionGenericsOrdered[paramName] = GenericType(
+                    paramName,
+                    typeBound = bound,
+                    constraints = constraints,
+                    context = context
+                )
+            } else {
+                functionGenericsOrdered[paramName]?.constraints?.addAll(constraints)
+            }
+        }
+
+        functionGenericsOrdered.forEach { it.value?.let { it1 -> funGenerics.add(it1) } }
+        return funGenerics
     }
 }
